Now adding pieces to the board on click (PP)
Right now we are adding a cyan colored checkers piece to
any square that is clicked. The piece currently has no
marker for which player it could belogn to, or any
movement options. Current implementation is purely just
to see how the composition works and how button clicks
change state.

Next steps are to fix weird spacing issue with the board
and to put the pieces on the right starting spot, and to
give them player identifiers.