Start of piece logic migration
Have the skeleton of piece logic now in `PieceHandler.js`. Pawns are
back to where they were, working on the rook logic at the moment. Want
to break out the logic further from the original, where a function is
responsible for 'running' either a row, column or diagnol that a piece
could move on. This change should help reduce the repetition in the
original implementation of the rook piece.